client/requirements.txt
grpcio-tools==1.41.0
flask==3.0.0
Flask-Cors==4.0.0